alloy xml control is added for commands.

+++ Source/eu.modelwriter.marker.command/src/eu/modelwriter/marker/command/DeleteHandler.java
@@ -54,10 +54,8 @@ public class DeleteHandler extends AbstractHandler {
   private ArrayList<IMarker> candidateToTypeChanging;
 
   private void deleteFromAlloyXML(IMarker beDeleted) {
-    if (AlloyUtilities.isExists()) {
-      AlloyUtilities.removeMarker(beDeleted);
-      AlloyUtilities.removeRelationOfMarker(beDeleted);
-    }
+    AlloyUtilities.removeMarker(beDeleted);
+    AlloyUtilities.removeRelationOfMarker(beDeleted);
   }
 
   private void deleteMarker() {
@@ -109,15 +107,22 @@ public class DeleteHandler extends AbstractHandler {
 
   @Override
   public Object execute(ExecutionEvent event) throws ExecutionException {
-    this.editor =
-        PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage().getActiveEditor();
-    this.file = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage()
-        .getActiveEditor().getEditorInput().getAdapter(IFile.class);
-    this.selection =
-        PlatformUI.getWorkbench().getActiveWorkbenchWindow().getSelectionService().getSelection();
-    this.candidateToTypeChanging = new ArrayList<IMarker>();
-    deleteMarker();
-    refresh();
+    if (AlloyUtilities.isExists()) {
+      this.editor =
+          PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage().getActiveEditor();
+      this.file = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage()
+          .getActiveEditor().getEditorInput().getAdapter(IFile.class);
+      this.selection =
+          PlatformUI.getWorkbench().getActiveWorkbenchWindow().getSelectionService().getSelection();
+      this.candidateToTypeChanging = new ArrayList<IMarker>();
+      deleteMarker();
+      refresh();
+    } else {
+      MessageDialog infoDialog = new MessageDialog(MarkerActivator.getShell(), "System Information",
+          null, "You dont have any registered alloy file to system.", MessageDialog.INFORMATION,
+          new String[] {"OK"}, 0);
+      infoDialog.open();
+    }
     return null;
   }
